pyproject.toml
[project]
name = "simple-eplus-drm-dl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "pywidevine>=1.9.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "nuitka>=2.8.9",
]
